vuln_scan/requirements.txt
Flask==3.1.1
requests==2.32.4
gunicorn==23.0.0
python-dotenv==1.0.0
tree-sitter==0.20.4
rich==13.7.0
json_repair==0.54.2
